Move molzip rgroup row from RGroupDecomp to ChemTransforms
