+++ Code/GraphMol/ChemTransforms/MolFragmenter.h
@@ -142,5 +142,19 @@ RDKIT_CHEMTRANSFORMS_EXPORT std::unique_ptr<ROMol> molzip(
 RDKIT_CHEMTRANSFORMS_EXPORT std::unique_ptr<ROMol> molzip(
     std::vector<ROMOL_SPTR> &decomposition,
     const MolzipParams &params = MolzipParams());
+
+//! \brief Molzip an RGroupRow back into the original molecule if possible
+/*!  This correctly handles broken cycles that can happend during arbitrary
+ *  RGroup Decomposition.
+ *
+ * @param row - rgroup row as returned by the rgroup decompisition
+ *
+ * optional:
+ * @param params - molzip parameters
+ *
+ * @return - the zipped molecule
+ */
+RDKIT_RGROUPDECOMPOSITION_EXPORT std::unique_ptr<ROMol> molzip(const std::map<std::string, ROMOL_SPTR> &row,
+                                                               const MolzipParams &params=MolzipParams());
 }  // namespace RDKit
 #endif
